refactor(nils-common): extract secret redaction to shared module
- Add `nils_common::redact` with `redact_text`, `RedactedString`, and
  named secret regexes; cover assignment and bare-token shapes with tests.
- Replace identical private copies in `web-evidence` and
  `agent-workflow-primitives::test_first_evidence` with the shared helpers
  and drop the now-unused `regex` dep from `web-evidence`.
