+++ crates/nils-common/src/lib.rs
@@ -31,6 +31,7 @@ pub mod markdown;
 pub mod process;
 pub mod provider_runtime;
 pub mod rate_limits_ansi;
+pub mod redact;
 pub mod shell;
 
 pub fn greeting(name: &str) -> String {
