+++ crates/web-evidence/Cargo.toml
@@ -19,7 +19,6 @@ path = "src/main.rs"
 clap = { workspace = true }
 clap_complete = { workspace = true }
 nils-common = { version = "0.22.3", path = "../nils-common", package = "nils-common" }
-regex = "1"
 reqwest = { version = "0.13", default-features = false, features = ["blocking", "rustls"] }
 serde = { workspace = true }
 serde_json = { workspace = true }
