+++ crates/web-evidence/src/lib.rs
@@ -6,12 +6,10 @@ use std::ffi::OsString;
 use std::fs;
 use std::io::Read;
 use std::path::{Component, Path, PathBuf};
-use std::sync::OnceLock;
 use std::time::{Duration, SystemTime, UNIX_EPOCH};
 
 use clap::Parser;
 use clap::error::ErrorKind;
-use regex::Regex;
 use reqwest::Url;
 use reqwest::blocking::Client;
 use reqwest::header::{ACCEPT, HeaderMap, USER_AGENT};
@@ -20,6 +18,7 @@ use serde_json::{Value, json};
 
 use cli::{CaptureArgs, Cli, Command, HttpMethod, OutputFormat};
 use nils_common::cli_contract::exit;
+use nils_common::redact::{RedactedString, redact_text};
 
 const EXIT_OK: i32 = exit::SUCCESS;
 const EXIT_RUNTIME: i32 = exit::RUNTIME;
@@ -643,64 +642,6 @@ fn redact_url(url: &Url) -> RedactedString {
     }
 }
 
-fn redact_text(input: &str) -> RedactedString {
-    let mut value = input.to_string();
-    let mut replacements = 0usize;
-
-    let assignment_replacements = assignment_secret_regex().captures_iter(&value).count();
-    if assignment_replacements > 0 {
-        value = assignment_secret_regex()
-            .replace_all(&value, "${key}${after_key}[REDACTED]")
-            .to_string();
-        replacements += assignment_replacements;
-    }
-
-    let token_replacements = token_secret_regex().find_iter(&value).count();
-    if token_replacements > 0 {
-        value = token_secret_regex()
-            .replace_all(&value, "[REDACTED]")
-            .to_string();
-        replacements += token_replacements;
-    }
-
-    RedactedString {
-        value,
-        replacements,
-    }
-}
-
-fn assignment_secret_regex() -> &'static Regex {
-    static RE: OnceLock<Regex> = OnceLock::new();
-    RE.get_or_init(|| {
-        Regex::new(
-            r#"(?ix)
-            (?P<key>\b(?:access[_-]?token|refresh[_-]?token|api[_-]?key|apikey|authorization|cookie|password|secret|session[_-]?id|token)\b)
-            (?P<after_key>"?\s*[:=]\s*)
-            (?P<value>"[^"]*"|'[^']*'|[^\s,;&}\]]+)
-            "#,
-        )
-        .expect("valid assignment secret regex")
-    })
-}
-
-fn token_secret_regex() -> &'static Regex {
-    static RE: OnceLock<Regex> = OnceLock::new();
-    RE.get_or_init(|| {
-        Regex::new(
-            r#"(?ix)
-            \b(
-                sk-(?:proj-)?[A-Za-z0-9_-]{8,}
-                | ghp_[A-Za-z0-9_]{8,}
-                | github_pat_[A-Za-z0-9_]{8,}
-                | xox[baprs]-[A-Za-z0-9-]{8,}
-                | bearer\s+[A-Za-z0-9._~+/=-]{8,}
-            )\b
-            "#,
-        )
-        .expect("valid token secret regex")
-    })
-}
-
 fn is_sensitive_key(key: &str) -> bool {
     let lower = key.to_ascii_lowercase();
     lower.contains("token")
@@ -838,12 +779,6 @@ struct CaptureOutcome {
 struct BodyArtifact {
     ref_: ArtifactRef,
     redaction_replacements: usize,
-}
-
-#[derive(Debug)]
-struct RedactedString {
-    value: String,
-    replacements: usize,
 }
 
 #[derive(Debug)]
